# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * GymLink API specification
 * OpenAPI spec version: 0.1.0
 */
export interface ReportBody {
  /**
   * @minLength 1
   * @maxLength 120
   */
  reason: string;
  /** @maxLength 2000 */
  details?: string;
}

export interface BlockedUser {
  blockedId: string;
  name: string;
  avatar: string;
  avatarUrl?: string | null;
  createdAt: string;
}

export interface HealthStatus {
  status: string;
}

export type ConnectionType =
  (typeof ConnectionType)[keyof typeof ConnectionType];

export const ConnectionType = {
  crush: "crush",
  buddy: "buddy",
  advisor: "advisor",
  spotter: "spotter",
} as const;

export type ConnectionStatus =
  (typeof ConnectionStatus)[keyof typeof ConnectionStatus];

export const ConnectionStatus = {
  pending: "pending",
  accepted: "accepted",
  declined: "declined",
} as const;

export interface Gym {
  id: string;
  name: string;
  address: string;
  city: string;
  memberCount: number;
}

export interface User {
  id: string;
  name: string;
  age: number;
  /** Emoji or image URL */
  avatar: string;
  bio: string;
  gym: string;
  gymId?: string | null;
  schedule: string;
  interests: string[];
  verified: boolean;
  distance?: string;
  isMe?: boolean;
  activeNow?: boolean;
  checkedIn?: boolean;
  avatarUrl?: string | null;
  hidden?: boolean;
}

export interface UpdateProfileBody {
  name?: string;
  age?: number;
  bio?: string;
  gym?: string;
  gymId?: string;
  schedule?: string;
  interests?: string[];
  avatarUrl?: string | null;
  hidden?: boolean;
}

export interface CheckInBody {
  gymId: string;
  gymName: string;
}

export interface Connection {
  id: string;
  fromUserId: string;
  toUserId: string;
  type: ConnectionType;
  status: ConnectionStatus;
  anonymous: boolean;
  createdAt: string;
  fromUser?: User;
  toUser?: User;
}

export interface CreateConnectionBody {
  toUserId: string;
  type: ConnectionType;
  anonymous: boolean;
  mutualNotify?: boolean;
}

export type RespondConnectionBodyResponse =
  (typeof RespondConnectionBodyResponse)[keyof typeof RespondConnectionBodyResponse];

export const RespondConnectionBodyResponse = {
  accept: "accept",
  decline: "decline",
} as const;

export interface RespondConnectionBody {
  response: RespondConnectionBodyResponse;
}

export interface Notification {
  id: string;
  connectionId: string;
  type: ConnectionType;
  fromName: string;
  anonymous: boolean;
  read: boolean;
  responded: boolean;
  createdAt: string;
}

export interface GymStats {
  activeNow: number;
  totalMembers: number;
  crushCount: number;
  buddyCount: number;
  advisorCount: number;
  spotterCount: number;
  gymName: string;
}

export interface UploadUrlRequest {
  name: string;
  size: number;
  contentType: string;
}

export interface UploadUrlResponse {
  uploadURL: string;
  objectPath: string;
}

export interface WorkoutVideo {
  id: string;
  userId: string;
  objectPath: string;
  title: string;
  description?: string | null;
  createdAt: string;
  likeCount: number;
  likedByMe: boolean;
}

export interface CreateWorkoutVideoBody {
  objectPath: string;
  title: string;
  description?: string;
}

export interface UpdateWorkoutVideoBody {
  title: string;
}

export interface VideoLikeStatus {
  videoId: string;
  likeCount: number;
  likedByMe: boolean;
}

export type ListUsersParams = {
  gym?: string;
  interest?: string;
  search?: string;
};

export type ListConnectionsParams = {
  type?: ConnectionType;
  status?: ConnectionStatus;
};

export type ListWorkoutVideosParams = {
  userId?: string;
};

export type BlockUser201 = {
  ok?: boolean;
};

export type ReportUser201 = {
  ok?: boolean;
};

export type ReportVideo201 = {
  ok?: boolean;
};