# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: GymLink API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /connections:
    post:
      operationId: createConnection
      tags: [connections]
      summary: Send a connection request
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateConnectionBody"
      responses:
        "201":
          description: Connection created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Connection"
    get:
      operationId: listConnections
      tags: [connections]
      summary: List all connections for current user
      parameters:
        - name: type
          in: query
          schema:
            $ref: "#/components/schemas/ConnectionType"
        - name: status
          in: query
          schema:
            $ref: "#/components/schemas/ConnectionStatus"
      responses:
        "200":
          description: List of connections
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Connection"
  /connections/{id}/respond:
    post:
      operationId: respondToConnection
      tags: [connections]
      summary: Accept or decline a connection
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RespondConnectionBody"
      responses:
        "200":
          description: Connection updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Connection"
...
  /notifications/{id}/read:
    post:
      operationId: markNotificationRead
      tags: [notifications]
      summary: Mark a notification as read
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Notification marked read
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Notification"
...
  /users/{id}/block:
    post:
      operationId: blockUser
      tags: [users]
      summary: Block a user
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "201":
          description: Blocked
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
    delete:
      operationId: unblockUser
      tags: [users]
      summary: Unblock a user
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Unblocked
  /users/{id}/report:
    post:
      operationId: reportUser
      tags: [users]
      summary: Report a user for inappropriate content or behavior
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReportBody"
      responses:
        "201":
          description: Reported
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
  /videos/{id}/report:
    post:
      operationId: reportVideo
      tags: [videos]
      summary: Report a video for inappropriate content
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReportBody"
      responses:
        "201":
          description: Reported
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
...
components:
  schemas:
    ReportBody:
      type: object
      required: [reason]
      properties:
        reason:
          type: string
          minLength: 1
          maxLength: 120
        details:
          type: string
          maxLength: 2000
...
    ConnectionType:
      type: string
      enum: [crush, buddy, advisor, spotter]
    ConnectionStatus:
      type: string
      enum: [pending, accepted, declined]
...
    User:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        age:
          type: integer
        avatar:
          type: string
          description: Emoji or image URL
        bio:
          type: string
        gym:
          type: string
        gymId:
          type: string
          nullable: true
        schedule:
          type: string
        interests:
          type: array
          items:
            type: string
        verified:
          type: boolean
        distance:
          type: string
        isMe:
          type: boolean
        activeNow:
          type: boolean
        checkedIn:
          type: boolean
        avatarUrl:
          type: string
          nullable: true
        hidden:
          type: boolean
      required:
        - id
        - name
        - age
        - avatar
        - bio
        - gym
        - schedule
        - interests
        - verified
...
    Connection:
      type: object
      properties:
        id:
          type: string
        fromUserId:
          type: string
        toUserId:
          type: string
        type:
          $ref: "#/components/schemas/ConnectionType"
        status:
          $ref: "#/components/schemas/ConnectionStatus"
        anonymous:
          type: boolean
        createdAt:
          type: string
          format: date-time
        fromUser:
          $ref: "#/components/schemas/User"
        toUser:
          $ref: "#/components/schemas/User"
      required:
        - id
        - fromUserId
        - toUserId
        - type
        - status
        - anonymous
        - createdAt
    CreateConnectionBody:
      type: object
      properties:
        toUserId:
          type: string
        type:
          $ref: "#/components/schemas/ConnectionType"
        anonymous:
          type: boolean
        mutualNotify:
          type: boolean
      required:
        - toUserId
        - type
        - anonymous
    RespondConnectionBody:
      type: object
      properties:
        response:
          type: string
          enum: [accept, decline]
      required:
        - response
    Notification:
      type: object
      properties:
        id:
          type: string
        connectionId:
          type: string
        type:
          $ref: "#/components/schemas/ConnectionType"
        fromName:
          type: string
        anonymous:
          type: boolean
        read:
          type: boolean
        responded:
          type: boolean
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - connectionId
        - type
        - fromName
        - anonymous
        - read
        - responded
        - createdAt
...